[rename] 디렉토리를 수정한다.
