+++ data/src/main/java/com/example/data/remote/apirul/ApiUrl.kt
@@ -1,4 +1,4 @@
-package com.example.data.network
+package com.example.data.remote.apirul
 
 class ApiUrl {
     val book = "/book"
