+++ data/src/main/java/com/example/data/remote/config/NetworkConfig.kt
@@ -1,4 +1,4 @@
-package com.example.data.network
+package com.example.data.remote.config
 
 object NetworkConfig {
     val BaseUrl = "https://openapi.naver.com/v1/datalab/search"
