+++ data/src/main/java/com/example/data/remote/network/NetWorkInterceptor.kt
@@ -1,4 +1,4 @@
-package com.example.data.network
+package com.example.data.remote.network
 
 import okhttp3.Interceptor
 import okhttp3.Response
